// src/app/shared/model/slide.model.ts
/**
 * © Rolling Array https://rollingarray.co.in/
 *
 *
 * @summary Slide model
 * @author code@rollingarray.co.in
 *
 * Created at     : 2021-12-26 20:35:14 
 * Last modified  : 2021-12-26 20:35:48
 */


export interface SlideModel
{
  title: string;
  info: string;
  imageName: string;
  buttonText: string;
  helpUrl: string;
}